@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Palmtree\Html\Test;
 
 use Palmtree\Html\Element;
@@ -7,14 +9,14 @@ use PHPUnit\Framework\TestCase;
 
 class ElementTest extends TestCase
 {
-    public function testRender()
+    public function testRender(): void
     {
         $element = Element::create('div')->addChild(Element::create('span'));
 
         $this->assertSame("<div>\n    <span></span>\n</div>", $element->render());
     }
 
-    public function testTabSize()
+    public function testTabSize(): void
     {
         $element = Element::create('div')
             ->addChild(Element::create('span'))
@@ -23,7 +25,7 @@ class ElementTest extends TestCase
         $this->assertSame("<div>\n  <span></span>\n</div>", $element->render());
     }
 
-    public function testUseTab()
+    public function testUseTab(): void
     {
         $element = Element::create('div')
             ->addChild(Element::create('span'))
@@ -32,7 +34,7 @@ class ElementTest extends TestCase
         $this->assertSame("<div>\n\t<span></span>\n</div>", $element->render());
     }
 
-    public function testAddClass()
+    public function testAddClass(): void
     {
         $element = Element::create('div');
         $element->classes->add('foo', 'bar');
@@ -41,36 +43,41 @@ class ElementTest extends TestCase
         $this->assertContains('bar', $element->classes);
     }
 
-    public function testAttributes()
+    public function testAttributes(): void
     {
         $element = Element::create('input#foo.bar.baz');
 
         $element->attributes
             ->setData('foo', 'bar')
             ->set('type', 'checkbox')
-            ->set('checked', true)
+            ->set('checked')
         ;
 
-        $document = new \DOMDocument();
-        $document->loadHTML($element->render());
-        $node = $document->getElementsByTagName('body')->item(0)->childNodes->item(0);
+        $html = $element->render();
 
+        $document = new \DOMDocument();
+        $document->loadHTML($html);
+        $node = $document->getElementsByTagName('body')->item(0)?->childNodes->item(0);
+
+        $this->assertInstanceOf(\DOMElement::class, $node);
         $this->assertSame('input', $node->tagName);
         $this->assertSame('checkbox', $node->getAttribute('type'));
         $this->assertSame('foo', $node->getAttribute('id'));
         $this->assertSame('bar baz', $node->getAttribute('class'));
         $this->assertSame('bar', $node->getAttribute('data-foo'));
-        $this->assertSame('1', $node->getAttribute('checked'));
+        $this->assertTrue($node->hasAttribute('checked'));
+
+        $this->assertSame('<input class="bar baz" id="foo" data-foo="bar" type="checkbox" checked>' . \PHP_EOL, $html);
     }
 
-    public function testAttributeSelector()
+    public function testAttributeSelector(): void
     {
         $element = Element::create('input[type=checkbox][checked]');
 
-        $this->assertSame('<input type="checkbox" checked>' . PHP_EOL, $element->render());
+        $this->assertSame('<input type="checkbox" checked>' . \PHP_EOL, $element->render());
     }
 
-    public function testInnerText()
+    public function testInnerText(): void
     {
         $div = Element::create('div')->setInnerText('Hello, World!');
 
@@ -79,7 +86,7 @@ class ElementTest extends TestCase
         $this->assertSame('<div>Hello, World!</div>', $div->render());
     }
 
-    public function testDefaultGetters()
+    public function testDefaultGetters(): void
     {
         $div = new Element('div');
 
